feat: move bridges filters up
